Allow Nagios objects to override NagVis attributes
Allows NagVis object attributes to be overwritten by attributes
defined as Nagios objects.

This can be helpful e. g. when you want to override the host_url of
hosts in a dynmap (i. e. the link destination of a host). By adding a
custom variable '_url' in the Nagios host (template) definition you can
override the NagVis object attribute 'url'. This also works for any
other attribute.

The feature can be enabled by setting merge_custom_vars=1 in the
defaults section. Of course this feature has the ability to break things
if not used carefully.